Relocating all attribute section components to product-editor package
