+++ packages/js/product-editor/package.json
@@ -35,7 +35,7 @@
 		"@woocommerce/components": "workspace:*",
 		"@woocommerce/currency": "workspace:*",
 		"@woocommerce/customer-effort-score": "workspace:*",
-		"@woocommerce/data": "workspace:^4.1.0",
+		"@woocommerce/data": "workspace:*",
 		"@woocommerce/experimental": "workspace:*",
 		"@woocommerce/navigation": "workspace:^8.1.0",
 		"@woocommerce/number": "workspace:*",
@@ -89,6 +89,7 @@
 		"@types/wordpress__keycodes": "^2.3.1",
 		"@types/wordpress__media-utils": "^3.0.0",
 		"@types/wordpress__plugins": "^3.0.0",
+		"@types/wordpress__rich-text": "^3.4.6",
 		"@woocommerce/eslint-plugin": "workspace:*",
 		"@woocommerce/internal-js-tests": "workspace:*",
 		"@woocommerce/internal-style-build": "workspace:*",
