+++ packages/js/product-editor/src/components/attribute-control/edit-attribute-modal.scss
