+++ packages/js/product-editor/src/components/attribute-control/index.ts
