+++ packages/js/product-editor/src/components/attribute-control/new-attribute-modal.scss
