+++ packages/js/product-editor/src/components/attribute-control/test/new-attribute-modal.spec.tsx
