+++ packages/js/product-editor/src/components/attribute-control/utils.ts
