+++ packages/js/product-editor/src/components/attribute-empty-state/attribute-empty-state.scss
