+++ packages/js/product-editor/src/components/attribute-empty-state/index.ts
@@ -1,0 +1,1 @@
+export * from './attribute-empty-state';
