+++ packages/js/product-editor/src/components/attribute-input-field/attribute-input-field.scss
