+++ packages/js/product-editor/src/components/attribute-list-item/attribute-list-item.scss
