+++ packages/js/product-editor/src/components/attribute-list-item/index.ts
