+++ packages/js/product-editor/src/components/attribute-term-input-field/attribute-term-input-field.scss
