+++ packages/js/product-editor/src/components/attribute-term-input-field/create-attribute-term-modal.scss
