+++ packages/js/product-editor/src/components/attributes/index.ts
