+++ packages/js/product-editor/src/components/index.ts
@@ -27,3 +27,6 @@ export {
 	type ShippingDimensionsImageProps,
 	type HighlightSides,
 } from './shipping-dimensions-image';
+
+export { AttributeControl as __experimentalAttributeControl } from './attribute-control';
+export { Attributes as __experimentalAttributes } from './attributes';
