+++ packages/js/product-editor/src/style.scss
@@ -23,6 +23,14 @@
 @import 'components/edit-product-link-modal/style.scss';
 @import 'components/details-categories-field/style.scss';
 @import 'components/details-categories-field/create-category-modal.scss';
+@import 'components/attribute-control/attribute-field.scss';
+@import 'components/attribute-control/edit-attribute-modal.scss';
+@import 'components/attribute-control/new-attribute-modal.scss';
+@import 'components/attribute-input-field/attribute-input-field.scss';
+@import 'components/attribute-empty-state/attribute-empty-state.scss';
+@import 'components/attribute-list-item/attribute-list-item.scss';
+@import 'components/attribute-term-input-field/attribute-term-input-field.scss';
+@import 'components/attribute-term-input-field/create-attribute-term-modal.scss';
 
 /* Field Blocks */
