--- plugins/woocommerce-admin/client/products/fields/attribute-empty-state/index.ts
@@ -1,2 +1,0 @@
-export * from './attribute-empty-state';
-export { default as AttributeEmptyStateLogo } from './attribute-empty-state-logo.svg';
